+++ R/make_units.R
@@ -103,10 +103,10 @@
 #' # For example, a microbiologist might work with counts of bacterial cells
 #' # make_units(cells/ml) # by default, throws an ERROR
 #' # First define the unit, then the newly defined unit is accepted.
-#' define_new_symbolic_unit("cells")
+#' install_symbolic_unit("cells")
 #' make_units(cells/ml) 
 #' 
-#' # Note, define_new_symbolic_unit() does not add any support for unit
+#' # Note, install_symbolic_unit() does not add any support for unit
 #' # converstion, or arathmetic operations that require unit conversion. See
 #' # ?install_conversion_function for how to define relationships for user 
 #' # defined units.
@@ -292,7 +292,7 @@ pc <- function(x) {
   paste0("In ", sQuote(full_expr), ", ", 
     pc(sQuote(unrecognized_symbols)), " ", is_are, " not recognized by udunits.\n",
     "See a table of valid unit symbols and names with valid_udunits().\n", 
-    "Add user-defined units with define_new_symbolic_unit().")
+    "Add user-defined units with install_symbolic_unit().")
 }
 
 is_valid_unit_symbol <- function(chr) {
@@ -309,7 +309,7 @@ is_valid_unit_symbol <- function(chr) {
 #'   individually, such that each symbol must be recognized by the udunits
 #'   database (checked by \code{ud.is.parseable()}, \emph{or} be a custom,
 #'   user-defined unit symbol that was defined either by
-#'   \code{define_new_symbolic_unit()} or \code{install_conversion_function()}.
+#'   \code{install_symbolic_unit()} or \code{install_conversion_function()}.
 #'   To see which symbols and names are currently recognized by the udunits
 #'   database, see \code{udunits_symbols()}.
 #'